move templating stuff into its own package under core
